refactored the search test into its own file to maintain the existing structure of the codebase
